Move reporting.model to dto.
