+++ src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/cas2hdc/dto/reference/Cas2PersistedApplicationStatus.kt
@@ -1,4 +1,4 @@
-package uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2hdc.reporting.model.reference
+package uk.gov.justice.digital.hmpps.approvedpremisesapi.cas2hdc.dto.reference
 
 import java.util.UUID
